automap supports for decorator/options
automap now supports, for proper function func :

  func(...)
  decorator(func)(...)
  func(opt,...)
